refactor: shrink AccountPosition + EModeCategory storage; bps fields → u32
AccountPosition drops account_id/position_type/asset (derivable from storage
key + map key), bps → u32. ~33% smaller per stored position.

EModeCategory absorbs the per-category asset map (was a second ledger entry)
and drops category_id (storage-key discriminant). One ledger entry per
category instead of two.

bps fields across MarketParams / InterestRateModel / AssetConfig /
OraclePriceFluctuation / event payloads → u32; Bps/Ray/Wad::from_raw now
polymorphic over impl Into<i128> so storage and i128 constants share one
constructor. Validators (validate_and_calculate_tolerances, is_within_anchor,
add/edit_e_mode_category, edit_oracle_tolerance) take u32 directly.

Pool ABI: seize_position adds a side: AccountPositionType arg; update_params
takes reserve_factor: u32; the require_same_asset defense was tied to the
removed position.asset and goes away.

View get_account_positions now returns the per-side maps directly so SDK
consumers receive the asset alongside each snapshot.

Event payloads (EventAccountPosition, EventEModeCategory) re-inject the
fields the storage no longer carries — indexer schemas stay stable.

cargo check / clippy / test (59 suites): clean.

+++ test-harness/src/admin.rs
@@ -102,7 +102,7 @@ impl LendingTest {
     // E-mode management
     // -----------------------------------------------------------------------
 
-    pub fn edit_e_mode_category(&self, category_id: u32, ltv: i128, threshold: i128, bonus: i128) {
+    pub fn edit_e_mode_category(&self, category_id: u32, ltv: u32, threshold: u32, bonus: u32) {
         self.ctrl_client()
             .edit_e_mode_category(&category_id, &ltv, &threshold, &bonus);
     }
